Update decryptor to changes in gensim
Gensim has changed and suddenly the "KeyedVector" class lacked a "wv" subclass. Also the "vocab" method was removed/replaced. All this was poorly documented on the gensim 4 migration guide, which is contradicting itself by yet claiming the existence of a "wv" subclass.
Fortunatly, I was able to find a new way to call the "similarity" and most similar methods and a method which replaces the functionality of the old "vocab" method, to makes this game run again unter pathon 3.7 and gensim 4.xx.